requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
torch>=2.4.0
transformers>=4.35.0
scikit-learn>=1.3.0
huggingface-hub>=0.19.0
ultralytics>=8.0.200
opencv-python>=4.8.0
easyocr>=1.7.0
pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
pdfplumber>=0.9.0
PyPDF2>=3.0.0
python-docx>=0.8.11
python-pptx>=0.6.21
pdf2image>=1.16.0
nomic>=1.1.0
python-dateutil>=2.8.0
requests>=2.28.0